Фиксы после ревью - добавил loading view для экрана было/стало - PermissionHandler - добавил статус `delivery` - переименовал некоторые dto

+++ feature-route/impl/src/client/kotlin/root/domain/model/NewOrderModel.kt
@@ -2,7 +2,7 @@ package root.domain.model
 
 @Suppress("LongParameterList")
 class NewOrderModel(
-    val addressId: Int,
+    val addressId: Long,
     val arrivalDay: String,
     val arrivalTime: String,
     val cargoType: String,
